requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.0
openai==1.59.6
twilio==8.10.0
pydub==0.25.1
sentence-transformers==3.4.1
scikit-learn==1.3.2
PyPDF2==3.0.1
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.26.2
smallestai==2.0.0
python-jose==3.3.0
pydantic_settings==2.7.1
huggingface_hub==0.28.1
websockets==14.2
openai-whisper==20240930 #3.0.2